tenra_web/backend/requirements.txt
fastapi
uvicorn
requests
pywebview
pillow
pyautogui
keyboard
send2trash
